requirements.txt
drf_autodocs==0.4.3
adal==0.4.7
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.23.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
azure==2.0.0
azure-batch==3.0.0
azure-common==1.1.8
azure-datalake-store==0.0.17
azure-graphrbac==0.30.0
azure-keyvault==0.3.7
azure-mgmt==1.0.0
azure-mgmt-authorization==0.30.0
azure-mgmt-batch==4.0.0
azure-mgmt-cdn==0.30.3
azure-mgmt-cognitiveservices==1.0.0
azure-mgmt-compute==1.0.0
azure-mgmt-containerregistry==0.2.1
azure-mgmt-datalake-analytics==0.1.6
azure-mgmt-datalake-nspkg==2.0.0
azure-mgmt-datalake-store==0.1.6
azure-mgmt-devtestlabs==2.0.0
azure-mgmt-dns==1.0.1
azure-mgmt-documentdb==0.1.3
azure-mgmt-iothub==0.2.2
azure-mgmt-keyvault==0.31.0
azure-mgmt-logic==2.1.0
azure-mgmt-monitor==0.2.1
azure-mgmt-network==1.0.0
azure-mgmt-nspkg==2.0.0
azure-mgmt-rdbms==0.1.0
azure-mgmt-redis==4.1.0
azure-mgmt-resource==1.1.0
azure-mgmt-scheduler==1.1.3
azure-mgmt-sql==0.5.3
azure-mgmt-storage==1.0.0
azure-mgmt-trafficmanager==0.30.0
azure-mgmt-web==0.32.0
azure-nspkg==2.0.0
azure-servicebus==0.21.1
azure-servicefabric==5.6.130
azure-servicemanagement-legacy==0.20.6
azure-storage==0.20.0
boto3==1.7.69
botocore==1.10.69
certifi==2018.8.24
cffi==1.11.2
channels==1.1.8
channels-api==0.4.0
chardet==3.0.4
click==6.7
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.1
daphne==1.4.2
Django==1.11.15
django-appconf==1.0.2
django-braces==1.11.0
django-cors-headers==2.1.0
django-countries==5.2
django-filter==1.1.0
django-imagekit==4.0.1
django-oauth-toolkit==1.0.0
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.0.8
django-rest-swagger==2.1.2
django-s3-storage==0.12.4
django-sql-explorer==1.1.1
django-storages==1.6.5
djangorestframework==3.8.2
docutils==0.14
drf-openapi==1.3.0
drf-yasg==1.10.2
enum34==1.1.6
flex==6.13.2
futures==3.1.1
gunicorn==19.7.1
hyperlink==17.3.1
idna==2.7
incremental==17.5.0
inflection==0.3.1
ipaddress==1.0.18
isodate==0.6.0
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonpointer==1.14
jsonschema==2.6.0
keyring==10.4.0
MarkupSafe==1.0
msgpack-python==0.5.4
msrest==0.4.17
msrestazure==0.4.15
oauthlib==2.0.6
olefile==0.44
openapi-codec==1.3.2
pathlib2==2.3.0
pilkit==2.0
Pillow==4.3.0
psycopg2==2.7.3.1
pycparser==2.18
PyJWT==1.5.3
python-dateutil==2.6.1
python-openid==2.2.5
pytz==2018.5
PyYAML==3.13
redis==2.10.6
requests==2.19.1
requests-oauthlib==0.8.0
rfc3987==1.3.8
s3transfer==0.1.13
scandir==1.6
SecretStorage==2.3.1
simplejson==3.16.0
six==1.11.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
sqlparse==0.2.4
strict-rfc3339==0.7
swagger-spec-validator==2.4.0
Twisted==17.9.0
txaio==2.8.2
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
validate-email==1.3
whitenoise==3.3.1
zope.interface==4.4.3
